src/app/requirements.txt
streamlit
psycopg2-binary
pandas
sentence_transformers
pgvector
